pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "omniprompt"
version = "1.0.0"
description = "A command-line utility for quickly testing and interacting with various LLM APIs."
readme = "README.md"
authors = [
  { name = "Andrew Kamau" },
]
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "PyYAML",
    "google-genai",
    "openai",
    "anthropic",
    "groq",
    "dashscope",
    "rich",
    "requests",
]
requires-python = ">=3.9"

[project.urls]
"Homepage" = "https://github.com/andkamau/omniprompt"
"Bug Tracker" = "https://github.com/andkamau/omniprompt/issues"

[project.scripts]
omniprompt = "omniprompt.cli:main"
